+++ src/pages/posts-manager/model/NewComment.ts
@@ -1,0 +1,5 @@
+export interface NewComment {
+  body: string
+  postId: number | null
+  userId: number
+}
