+++ src/pages/posts-manager/model/Tag.ts
@@ -1,0 +1,5 @@
+export interface Tag {
+  slug: string
+  name: string
+  url: string
+}
