+++ src/pages/posts-manager/ui/AddPostDialog.tsx
@@ -1,0 +1,71 @@
+// 게시물 추가 대화상자
+
+import React from "react"
+import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, Input, Textarea } from "../../../shared/ui"
+import { Post } from "../model/Post"
+
+interface NewPost {
+  title: string
+  body: string
+  userId: number
+}
+
+interface Props {
+  showAddDialog: boolean
+  setShowAddDialog: React.Dispatch<React.SetStateAction<boolean>>
+  newPost: NewPost
+  setNewPost: React.Dispatch<React.SetStateAction<NewPost>>
+  setPosts: React.Dispatch<React.SetStateAction<Post[]>>
+  posts: Post[]
+}
+
+const AddPostDialog = ({ showAddDialog, setShowAddDialog, newPost, setNewPost, setPosts, posts }: Props) => {
+  // 게시물 추가
+  const addPost = async () => {
+    try {
+      const response = await fetch("/api/posts/add", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify(newPost),
+      })
+      const data = await response.json()
+      setPosts([data, ...posts])
+      setShowAddDialog(false)
+      setNewPost({ title: "", body: "", userId: 1 })
+    } catch (error) {
+      console.error("게시물 추가 오류:", error)
+    }
+  }
+
+  return (
+    <Dialog open={showAddDialog} onOpenChange={setShowAddDialog}>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>새 게시물 추가</DialogTitle>
+        </DialogHeader>
+        <div className="space-y-4">
+          <Input
+            placeholder="제목"
+            value={newPost.title}
+            onChange={(e) => setNewPost({ ...newPost, title: e.target.value })}
+          />
+          <Textarea
+            rows={30}
+            placeholder="내용"
+            value={newPost.body}
+            onChange={(e) => setNewPost({ ...newPost, body: e.target.value })}
+          />
+          <Input
+            type="number"
+            placeholder="사용자 ID"
+            value={newPost.userId}
+            onChange={(e) => setNewPost({ ...newPost, userId: Number(e.target.value) })}
+          />
+          <Button onClick={addPost}>게시물 추가</Button>
+        </div>
+      </DialogContent>
+    </Dialog>
+  )
+}
+
+export default AddPostDialog
