+++ src/pages/posts-manager/ui/PostDetailDialog.tsx
@@ -1,0 +1,133 @@
+import React from "react"
+import { Edit2, Plus, ThumbsUp, Trash2 } from "lucide-react"
+import { Button, Dialog, DialogContent, DialogHeader, DialogTitle } from "../../../shared/ui"
+import HighlightText from "../ui/HighlightText"
+import { NewComment } from "../model/NewComment"
+import { Comment, Comments } from "../model/Comment"
+import { Post } from "../model/Post"
+
+interface Props {
+  setNewComment: React.Dispatch<React.SetStateAction<NewComment>>
+  setShowAddCommentDialog: React.Dispatch<React.SetStateAction<boolean>>
+  comments: Comments
+  setSelectedComment: React.Dispatch<React.SetStateAction<Comment | null>>
+  setShowEditCommentDialog: React.Dispatch<React.SetStateAction<boolean>>
+  showPostDetailDialog: boolean
+  setShowPostDetailDialog: React.Dispatch<React.SetStateAction<boolean>>
+  selectedPost: Post | null
+  searchQuery: string
+  setComments: React.Dispatch<React.SetStateAction<Comments>>
+}
+
+const PostDetailDialog = ({
+  setNewComment,
+  setShowAddCommentDialog,
+  comments,
+  setSelectedComment,
+  setShowEditCommentDialog,
+  showPostDetailDialog,
+  setShowPostDetailDialog,
+  selectedPost,
+  searchQuery,
+  setComments,
+}: Props) => {
+  // 댓글 삭제
+  const deleteComment = async (id: number, postId: number) => {
+    try {
+      await fetch(`/api/comments/${id}`, {
+        method: "DELETE",
+      })
+      setComments((prev) => ({
+        ...prev,
+        [postId]: prev[postId].filter((comment) => comment.id !== id),
+      }))
+    } catch (error) {
+      console.error("댓글 삭제 오류:", error)
+    }
+  }
+  // 댓글 좋아요
+  const likeComment = async (id: number, postId: number) => {
+    try {
+      const response = await fetch(`/api/comments/${id}`, {
+        method: "PATCH",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ likes: comments[postId].find((c) => c.id === id)?.likes || 0 + 1 }),
+      })
+      const data = await response.json()
+      setComments((prev) => ({
+        ...prev,
+        [postId]: prev[postId].map((comment) => (comment.id === data.id ? data : comment)),
+      }))
+    } catch (error) {
+      console.error("댓글 좋아요 오류:", error)
+    }
+  }
+  // 댓글 렌더링
+  const renderComments = (postId: number) => (
+    <div className="mt-2">
+      <div className="flex items-center justify-between mb-2">
+        <h3 className="text-sm font-semibold">댓글</h3>
+        <Button
+          size="sm"
+          onClick={() => {
+            setNewComment((prev) => ({ ...prev, postId }))
+            setShowAddCommentDialog(true)
+          }}
+        >
+          <Plus className="w-3 h-3 mr-1" />
+          댓글 추가
+        </Button>
+      </div>
+      <div className="space-y-1">
+        {comments[postId]?.map((comment) => (
+          <div key={comment.id} className="flex items-center justify-between text-sm border-b pb-1">
+            <div className="flex items-center space-x-2 overflow-hidden">
+              <span className="font-medium truncate">{comment.user.username}:</span>
+              <span className="truncate">
+                <HighlightText text={comment.body} highlight={searchQuery} />
+              </span>
+            </div>
+            <div className="flex items-center space-x-1">
+              <Button variant="ghost" size="sm" onClick={() => likeComment(comment.id, postId)}>
+                <ThumbsUp className="w-3 h-3" />
+                <span className="ml-1 text-xs">{comment.likes}</span>
+              </Button>
+              <Button
+                variant="ghost"
+                size="sm"
+                onClick={() => {
+                  setSelectedComment(comment)
+                  setShowEditCommentDialog(true)
+                }}
+              >
+                <Edit2 className="w-3 h-3" />
+              </Button>
+              <Button variant="ghost" size="sm" onClick={() => deleteComment(comment.id, postId)}>
+                <Trash2 className="w-3 h-3" />
+              </Button>
+            </div>
+          </div>
+        ))}
+      </div>
+    </div>
+  )
+  return (
+    <Dialog open={showPostDetailDialog} onOpenChange={setShowPostDetailDialog}>
+      <DialogContent className="max-w-3xl">
+        <DialogHeader>
+          <DialogTitle>
+            <HighlightText text={selectedPost?.title} highlight={searchQuery} />
+          </DialogTitle>
+        </DialogHeader>
+        <div className="space-y-4">
+          <p>
+            <HighlightText text={selectedPost?.body} highlight={searchQuery} />
+          </p>
+          {renderComments(selectedPost?.id as number)}
+        </div>
+      </DialogContent>
+    </Dialog>
+  )
+}
+
+export default PostDetailDialog
