+++ src/pages/posts-manager/ui/PostsManagerHeader.tsx
@@ -1,0 +1,23 @@
+import React from "react"
+import { Button, CardHeader, CardTitle } from "../../../shared/ui"
+import { Plus } from "lucide-react"
+
+interface Props {
+  setShowAddDialog: React.Dispatch<React.SetStateAction<boolean>>
+}
+
+const PostsManagerHeader = ({ setShowAddDialog }: Props) => {
+  return (
+    <CardHeader>
+      <CardTitle className="flex items-center justify-between">
+        <span>게시물 관리자</span>
+        <Button onClick={() => setShowAddDialog(true)}>
+          <Plus className="w-4 h-4 mr-2" />
+          게시물 추가
+        </Button>
+      </CardTitle>
+    </CardHeader>
+  )
+}
+
+export default PostsManagerHeader
